+++ Mcode/+fancyclip/+internal/DispstrHelper.m
@@ -1,0 +1,16 @@
+classdef DispstrHelper
+  
+  methods (Static)
+    
+    function txt = disparray(x)
+      strs = dispstrs(x);
+      txt = dispstrlib.internal.DispstrImpl.prettyprintArray(strs);
+      if nargout == 0
+        disp(txt)
+        clear txt
+      end
+    end
+    
+  end
+  
+end