+++ Mcode/+fancyclip/+internal/LibraryInitializer.m
@@ -1,0 +1,12 @@
+classdef LibraryInitializer
+
+  methods
+    
+    function this = LibraryInitializer()
+      fancyclip.internal.initializePackage;
+    end
+    
+  end
+  
+end
+