+++ dev-kit/load_fancyclip.m
@@ -1,0 +1,5 @@
+function load_fancyclip
+
+repoDir = fileparts(fileparts(mfilename('fullpath')));
+toplevelDir = [repoDir '/Mcode'];
+addpath(toplevelDir);