+++ dev-kit/load_mypackage.m
@@ -1,5 +1,0 @@
-function load_mypackage
-
-repoDir = fileparts(fileparts(mfilename('fullpath')));
-toplevelDir = [repoDir '/Mcode'];
-addpath(toplevelDir);