+++ project_settings.sh
@@ -5,7 +5,7 @@
 
 # The name of your project and its GitHub repo. Capitalization should match what
 # your public "branding" is; this will show up in human-readable documentation.
-PROJECT=MyCoolProject
+PROJECT=fancyclip
 
 # The version of Matlab you're developing against. This will be the version of
 # Matlab that the project builds against (on Mac and Windows anyway), and the minimum
@@ -16,26 +16,26 @@ PROJECT_MATLAB_VERSION=R2019b
 # its code in. This is the "+<package>" directory that'll be directly under Mcode,
 # and is the "namespace" that your project lives in.
 # It is conventional for package names to be in all lower case.
-PACKAGE=mycoolpackage
+PACKAGE=fancyclip
 
 # Your GitHub user name or organization name that's hosting the project
-GHUSER=mygithubusername
+GHUSER=janklab
 
 # If you want to provide a contact email for your project, put it here. Optional.
-PROJECT_EMAIL=
+PROJECT_EMAIL=andrew@apjanke.net
 
 # The site generator tool you want to use for the project documentation.
 # Valid choices are "jekyll" and  "mkdocs".
 DOCSITETOOL=jekyll
 
 # Human-readable name of the project's primary author or maintainer
-PROJECT_AUTHOR="Your Name Here"
+PROJECT_AUTHOR="Andrew Janke"
 
 # Everything below here is optional! If you omit it, you'll end up with placeholder text
 # in some of your documentation, but the project will still work
 
 # One-sentence summary of the project. No <, >, or & characters allowed!
-PROJECT_SUMMARY="Short summary of project goes here"
+PROJECT_SUMMARY="Enhanced clipboard copy and paste for Matlab"
 
 # Multi-sentence project description. No <, >, or & characters allowed!
-PROJECT_DESCRIPTION="Longer description of project goes here."
+PROJECT_DESCRIPTION="Enhanced clipboard copy and paste for Matlab. Supports copying as HTML or JSON, pasting to Excel as tabular data, and more."